Move webpack client to separate module
